refactor(eslint): rename internal/eslint-plugin-xds dir to -astryx
Rename the internal ESLint plugin directory eslint-plugin-xds ->
eslint-plugin-astryx and update its path references:
- eslint.config.js import + file glob
- README self-reference import path
- pnpm-lock.yaml workspace importer key

The plugin's package name (@xds/eslint-plugin) and its '@xds' rule namespace
are unchanged — those belong to the package-scope rename tracked separately.
