+++ internal/eslint-plugin-astryx/boolean-prop-naming.js
