+++ internal/eslint-plugin-astryx/boolean-prop-naming.test.mjs
