+++ internal/eslint-plugin-astryx/copyright-header.js
