+++ internal/eslint-plugin-astryx/copyright-header.test.mjs
