+++ internal/eslint-plugin-astryx/no-border-shorthand.js
