+++ internal/eslint-plugin-astryx/no-border-shorthand.test.mjs
