+++ internal/eslint-plugin-astryx/no-classname-clobber.js
