+++ internal/eslint-plugin-astryx/no-hardcoded-anchor.js
