+++ internal/eslint-plugin-astryx/no-react-introspection.js
