+++ internal/eslint-plugin-astryx/no-react-introspection.test.mjs
