+++ internal/eslint-plugin-astryx/no-react-namespace-hooks.js
