+++ internal/eslint-plugin-astryx/no-react-namespace-hooks.test.mjs
