+++ internal/eslint-plugin-astryx/package.json
