+++ internal/eslint-plugin-astryx/presentational-component.js
