+++ internal/eslint-plugin-astryx/require-base-props.js
