+++ internal/eslint-plugin-astryx/require-ref-prop.js
