+++ internal/eslint-plugin-astryx/shared.js
